+++ src/Util/FileTransfer/OpencastIngestService.php
@@ -8,6 +8,8 @@ use srag\Plugins\Opencast\Model\Event\Request\UploadEventRequest;
 use srag\Plugins\Opencast\Util\Transformator\MetadataToXML;
 use xoctException;
 use srag\Plugins\Opencast\API\API;
+use ilFFmpeg;
+use ilShellUtil;
 
 class OpencastIngestService
 {
@@ -53,6 +55,41 @@ class OpencastIngestService
             $this->uploadStorageService->buildACLUploadFile($payload->getAcl())->getFileStream()
         );
 
+        // Subtitles using addTrack ingest method.
+        if ($payload->hasSubtitles()) {
+            foreach ($payload->getSubtitles() as $lang_code => $subtitle_uploadfile) {
+                // Perform conversion of supported subtitles files to WebVTT format.
+                $file_stream = $subtitle_uploadfile->getFileStream();
+                if (ilFFmpeg::enabled() && $subtitle_uploadfile->getMimeType() != 'text/vtt') {
+                    $path = $subtitle_uploadfile->getPath();
+                    $extension = pathinfo($path, PATHINFO_EXTENSION);
+                    $new_vtt_path = str_replace(".$extension", '.vtt', $path);
+                    $escaped_path = ilShellUtil::escapeShellArg($path);
+                    $escaped_new_vtt_path = ilShellUtil::escapeShellArg($new_vtt_path);
+                    $ffmpeg_cmd = "-i {$path} -c:s webvtt {$new_vtt_path}";
+                    $escaped_cmd = ilShellUtil::escapeShellCmd($ffmpeg_cmd);
+                    ilFFmpeg::exec($escaped_cmd);
+                    if (file_exists($new_vtt_path)) {
+                        $file_stream = fopen($new_vtt_path, 'rb');
+                        unlink($new_vtt_path);
+                    }
+                }
+                // Important tags to set for subtitles.
+                $tags = [
+                    'subtitle',
+                    'type:subtitle',
+                    'generator-type:manual',
+                    "lang:$lang_code",
+                ];
+                $media_package = $this->api->routes()->ingest->addTrack(
+                    $media_package,
+                    'captions/source', // Important flavor to set for subtitles.
+                    $file_stream,
+                    implode(',', $tags)
+                );
+            }
+        }
+
         // track
         $media_package = $this->api->routes()->ingest->addTrack(
             $media_package,
@@ -60,10 +97,14 @@ class OpencastIngestService
             $payload->getPresentation()->getFileStream()
         );
 
+        // Get workflow configuration params ready, make sure it is array!
+        $workflow_configuration = json_decode(json_encode($payload->getProcessing()->getConfiguration() ?? []), true);
         // ingest
         $media_package = $this->api->routes()->ingest->ingest(
             $media_package,
-            $payload->getProcessing()->getWorkflow()
+            $payload->getProcessing()->getWorkflow(),
+            '',
+            $workflow_configuration
         );
 
         // When we are done, we deactivate the ingest to keep everything clean.