+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+
+- [FEATURE] Upload Subtitles #10 (via PR #277)
 - [FEATURE] New Right management #14 (via PR #301)
 - [FIX] The Opencast object cannot be deleted #275
 - [FIX] make code compliant to the ILIAS 7 version